Desktop/IPtv_projects/Projects Eldad/Bot/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-dotenv==1.0.1
jinja2==3.1.3
pandas==2.2.2
